fixed GL* type names

+++ engine/animation/animation.h
@@ -11,11 +11,11 @@ public:
     vector<glm::vec3> locations;
     vector<glm::vec3> rotations;
     vector<glm::vec3> scales;
-    vector<uint> ids;
+    vector<size_t> ids;
     float poseTime;
     float(*timingFunction)(float) = [](float a) -> float {return a;};
 
-    void appendPosition(glm::vec3 location, glm::vec3 rotation, glm::vec3 scale, uint id){
+    void appendPosition(glm::vec3 location, glm::vec3 rotation, glm::vec3 scale, size_t id){
         locations.push_back(location);
         rotations.push_back(rotation);
         scales.push_back(scale);
